requirements.txt
pandas
python-dotenv
openai
google-generativeai
sentence-transformers
matplotlib
numpy
scikit-learn
statsmodels
